test(cypress): change setup for Bazel
